refactor: upgrade folders
