+++ package.json
@@ -65,6 +65,7 @@
     "eslint-import-resolver-typescript": "^3.5.2",
     "eslint-plugin-import": ">=2.25.3 <3.0.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-no-null": "^1.0.2",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.31.11",
     "eslint-plugin-react-hooks": "^4.6.0",
