--- src/core/constants/storage.ts
@@ -1,1 +1,0 @@
-export const STORAGE_REACTIONS: string = 'reactions';
