+++ src/core/contracts/homeScreens.ts
