+++ src/index.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './assets/index.css';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { Home } from '@/components/screens/home';
+import { HomePage } from '@/pages/home';
 
 const root: ReactDOM.Root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
   <React.StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/youtube-reload" element={<Home />} />
+        <Route path="/youtube-reload" element={<HomePage />} />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>,
